+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * Supported `NullValue` type as received type from `YDB`
 * Added `types.CastTo(types.Value, destination)` public method for cast `types.Value` to golang native type value destination
+* Added `types.TupleItem(types.Value)`, `types.StructFields(types.Value)` and `types.DictValues(types.Value)` funcs (extractors of internal fields of tuple, struct and dict values)
 * Refactored string representation of `types.Value`
 * Added supporting of `printf` formatter for `types.Value`